requirements.txt
anyio==4.8.0
APScheduler==3.6.3
asgiref==3.8.1
attrs==24.3.0
boto3==1.38.1
botocore==1.38.1
cachetools==4.2.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.6.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pandas==2.0.3
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2025.1
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.3.1
requests==2.32.3
rpds-py==0.22.3
s3transfer==0.12.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
telegram==0.0.1
tornado==6.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
